Ditch turicreate in favour of more common ML and NLP stuff
Turicreate was the library used by the MOOC I originally based this
code on, but Apple doesn't seem to have maintained it and it doesn't
work with current python versions so is a pain to use.

This version is now based on pandas, sklearn and nltk which are all
mainstream and open source packages.